doxygen comments and move MinHeapNode to separate class
